+++ modules/camera/camera_frame_sink.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "modules/camera/camera_frame.h"
+
+namespace cockpit {
+namespace camera {
+
+class CameraFrameSink {
+ public:
+  virtual ~CameraFrameSink() = default;
+
+  virtual bool Publish(CameraFrame frame) = 0;
+};
+
+}  // namespace camera
+}  // namespace cockpit